@@ -1,19 +1,22 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { Provider } from 'react-redux';
-import { configureStore } from './store.js';
-import App from './App.js';
-import { persistStore } from 'redux-persist';
-import { PersistGate } from 'redux-persist/lib/integration/react';
+import React from "react";
+import ReactDOM from "react-dom";
+import { Provider } from "react-redux";
+import { configureStore } from "./store.js";
+import App from "./App.js";
+import { persistStore } from "redux-persist";
+import { PersistGate } from "redux-persist/lib/integration/react";
+import { createRoot } from "react-dom/client";
 
 const store = configureStore();
 const persistor = persistStore(store);
 
-ReactDOM.render(
+const rootElement = document.getElementById("root");
+const root = createRoot(rootElement);
+
+root.render(
   <Provider store={store}>
     <PersistGate persistor={persistor} loading={<div>Loading...</div>}>
       <App />
     </PersistGate>
-  </Provider>,
-  document.getElementById('root')
+  </Provider>
 );
